Convert to .Net Standard
