+++ Sample/App.xaml
@@ -1,6 +1,7 @@
-﻿<Application x:Class="SharpAvi.Sample.App"
+﻿<Application x:Class="Sample.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="clr-namespace:Sample"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
