+++ Sample/Styles.xaml
@@ -1,21 +1,21 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:local="clr-namespace:SharpAvi.Sample">
+                    xmlns:local="clr-namespace:Sample">
 
     <BooleanToVisibilityConverter x:Key="VcBool2Vis"/>
     <local:BooleanToObjectConverter x:Key="VcBool2VisInv" FalseValue="Visible" TrueValue="Collapsed"/>
     <local:BooleanToObjectConverter x:Key="VcBoolInv" FalseValue="True" TrueValue="False"/>
-    
+
     <Style TargetType="TextBox">
         <Setter Property="BorderThickness" Value="2"/>
         <Setter Property="BorderBrush" Value="DarkGray"/>
         <Setter Property="Foreground" Value="DarkGray"/>
     </Style>
-    
+
     <Style TargetType="Label">
         <Setter Property="Foreground" Value="DarkGray"/>
     </Style>
-    
+
     <Style TargetType="Button">
         <Setter Property="MinWidth" Value="80"/>
         <Setter Property="Padding" Value="4,2"/>
@@ -84,7 +84,7 @@
         <Setter Property="Foreground" Value="DarkGray"/>
         <Setter Property="HorizontalContentAlignment" Value="Center"/>
         <Setter Property="VerticalContentAlignment" Value="Center"/>
-        
+
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="{x:Type ButtonBase}">
@@ -105,7 +105,7 @@
                 </ControlTemplate>
             </Setter.Value>
         </Setter>
-        
+
         <Style.Triggers>
             <Trigger Property="IsMouseOver" Value="True">
                 <Setter Property="BorderBrush" Value="DarkGray"/>
@@ -127,7 +127,7 @@
             </Trigger>
         </Style.Triggers>
     </Style>
-    
+
     <Style x:Key="RoundButtonGlyph" TargetType="{x:Type Shape}">
         <Setter Property="Fill" Value="{Binding Foreground, RelativeSource={RelativeSource AncestorType=ButtonBase}}"/>
         <Setter Property="StrokeThickness" Value="0"/>
@@ -138,9 +138,10 @@
             </Setter.Value>
         </Setter>
     </Style>
-    
+
     <Style TargetType="{x:Type Expander}">
         <Setter Property="BorderBrush" Value="LightGray"/>
         <Setter Property="Padding" Value="7"/>
     </Style>
+
 </ResourceDictionary>
