+++ Sample/WindowMoveBehavior.cs
@@ -1,11 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Windows;
 using System.Windows.Input;
 
-namespace SharpAvi.Sample
+namespace Sample
 {
     internal static class WindowMoveBehavior
     {
