--- Sample/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NAudio" version="1.7.1" targetFramework="net45" />
-</packages>
