--- SharpAvi.Fx35.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30723.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharpAvi", "SharpAvi\SharpAvi.Fx35.csproj", "{99ADB02B-C515-4F99-BA20-112CA82E2498}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sample", "Sample\Sample.Fx35.csproj", "{960B7709-DDB4-4F25-ADD8-936E9C218551}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{99ADB02B-C515-4F99-BA20-112CA82E2498}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{99ADB02B-C515-4F99-BA20-112CA82E2498}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{99ADB02B-C515-4F99-BA20-112CA82E2498}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{99ADB02B-C515-4F99-BA20-112CA82E2498}.Release|Any CPU.Build.0 = Release|Any CPU
-		{960B7709-DDB4-4F25-ADD8-936E9C218551}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{960B7709-DDB4-4F25-ADD8-936E9C218551}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{960B7709-DDB4-4F25-ADD8-936E9C218551}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{960B7709-DDB4-4F25-ADD8-936E9C218551}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
