--- SharpAvi.nuspec
@@ -1,28 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2011/08/nuspec.xsd">
-    <metadata>
-        <id>SharpAvi</id>
-        <version>2.1.2</version>
-        <authors>Vasili Maslov</authors>
-        <licenseUrl>https://github.com/baSSiLL/SharpAvi/blob/master/LICENSE.md</licenseUrl>
-        <projectUrl>https://github.com/baSSiLL/SharpAvi</projectUrl>
-        <requireLicenseAcceptance>false</requireLicenseAcceptance>
-        <description>If you want to render some video sequence, and do not want to touch DirectShow or depend on command-line utilities - SharpAvi may be what you need.
-
-Writing uncompressed AVI does not require any external dependencies, it's pure .NET code. Files are produced in compliance with the OpenDML extensions which allow (nearly) unlimited file size (no 2GB limit).
-
-Video is created by supplying individual in-memory bitmaps (byte arrays) and audio samples. Included are implementations of encoders for Motion JPEG (requires WPF), MPEG-4 (requires external VfW codecs) and MP3 (requires LAME binaries). Asynchronous writing/encoding is supported.</description>
-        <summary>A simple library for creating video files in AVI format.</summary>
-        <releaseNotes>Fixed uncompressed video for frame widths which are not multiples of 4.</releaseNotes>
-        <tags>AVI video authoring encoding OpenDML</tags>
-        <references>
-            <reference file="SharpAvi.dll" />
-        </references>
-    </metadata>
-    <files>
-        <file src="SharpAvi\bin\35\Release\SharpAvi.dll" target="lib\net35-client\SharpAvi.dll" />
-        <file src="SharpAvi\bin\35\Release\SharpAvi.XML" target="lib\net35-client\SharpAvi.XML" />
-        <file src="SharpAvi\bin\Release\SharpAvi.dll" target="lib\net45\SharpAvi.dll" />
-        <file src="SharpAvi\bin\Release\SharpAvi.XML" target="lib\net45\SharpAvi.XML" />
-    </files>
-</package>
