+++ SharpAvi/Codecs/BitmapUtils.cs
@@ -2,7 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+
+#if !NET35
 using System.Diagnostics.Contracts;
+#endif
 
 namespace SharpAvi.Codecs
 {
@@ -10,6 +13,7 @@ namespace SharpAvi.Codecs
     {
         public static unsafe void Bgr32ToBgr24(byte[] source, int srcOffset, byte[] destination, int destOffset, int pixelCount)
         {
+#if !NET35
             Contract.Requires(source != null);
             Contract.Requires(srcOffset >= 0);
             Contract.Requires(destination != null);
@@ -17,6 +21,7 @@ namespace SharpAvi.Codecs
             Contract.Requires(pixelCount >= 0);
             Contract.Requires(srcOffset + 4 * pixelCount <= source.Length);
             Contract.Requires(destOffset + 3 * pixelCount <= destination.Length);
+#endif
 
             fixed (byte* sourcePtr = source, destinationPtr = destination)
             {
@@ -37,12 +42,14 @@ namespace SharpAvi.Codecs
 
         public static void FlipVertical(byte[] source, int srcOffset, byte[] destination, int destOffset, int height, int stride)
         {
+#if !NET35
             Contract.Requires(source != null);
             Contract.Requires(destination != null);
             Contract.Requires(height >= 0);
             Contract.Requires(stride > 0);
             Contract.Requires(srcOffset + stride * height <= source.Length);
             Contract.Requires(destOffset + stride * height <= destination.Length);
+#endif
 
             var src = srcOffset;
             var dest = destOffset + (height - 1) * stride;
