+++ SharpAvi/Output/AsyncVideoStreamWrapper.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+#if !NET35
 using System.Diagnostics.Contracts;
+#endif
 using System.Linq;
 using System.Text;
-#if FX45
+#if !NET35
 using System.Threading.Tasks;
 #endif
 
@@ -19,7 +21,9 @@ namespace SharpAvi.Output
         public AsyncVideoStreamWrapper(IAviVideoStreamInternal baseStream)
             : base(baseStream)
         {
+#if !NET35
             Contract.Requires(baseStream != null);
+#endif
         }
 
         public override void WriteFrame(bool isKeyFrame, byte[] frameData, int startIndex, int length)
@@ -27,22 +31,22 @@ namespace SharpAvi.Output
             writeInvoker.Invoke(() => base.WriteFrame(isKeyFrame, frameData, startIndex, length));
         }
 
-#if FX45
-        public override Task WriteFrameAsync(bool isKeyFrame, byte[] frameData, int startIndex, int length)
-        {
-            return writeInvoker.InvokeAsync(() => base.WriteFrame(isKeyFrame, frameData, startIndex, length));
-        }
-#else
+#if NET35
         public override IAsyncResult BeginWriteFrame(bool isKeyFrame, byte[] frameData, int startIndex, int length, AsyncCallback userCallback, object stateObject)
         {
             return writeInvoker.BeginInvoke(
-                () => base.WriteFrame(isKeyFrame, frameData, startIndex, length), 
+                () => base.WriteFrame(isKeyFrame, frameData, startIndex, length),
                 userCallback, stateObject);
         }
 
         public override void EndWriteFrame(IAsyncResult asyncResult)
         {
             writeInvoker.EndInvoke(asyncResult);
+        }
+#else
+        public override Task WriteFrameAsync(bool isKeyFrame, byte[] frameData, int startIndex, int length)
+        {
+            return writeInvoker.InvokeAsync(() => base.WriteFrame(isKeyFrame, frameData, startIndex, length));
         }
 #endif
