+++ SharpAvi/Output/IAviAudioStream.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+#if !NET35
 using System.Diagnostics.Contracts;
+#endif
 using System.Linq;
 using System.Text;
-#if FX45
+#if !NET35
 using System.Threading.Tasks;
 #endif
 
@@ -12,7 +14,9 @@ namespace SharpAvi.Output
     /// <summary>
     /// Audio stream of AVI file.
     /// </summary>
+#if !NET35
     [ContractClass(typeof(Contracts.AviAudioStreamContract))]
+#endif
     public interface IAviAudioStream : IAviStream
     {
         /// <summary>
@@ -77,24 +81,7 @@ namespace SharpAvi.Output
         /// </remarks>
         void WriteBlock(byte[] data, int startIndex, int length);
 
-#if FX45
-        /// <summary>
-        /// Asynchronously writes a block of audio data.
-        /// </summary>
-        /// <param name="data">Data buffer.</param>
-        /// <param name="startIndex">Start index of data.</param>
-        /// <param name="length">Length of data.</param>
-        /// <returns>
-        /// A task representing the asynchronous write operation.
-        /// </returns>
-        /// <remarks>
-        /// Division of audio data into blocks may be arbitrary.
-        /// However, it is reasonable to write blocks of approximately the same duration
-        /// as a single video frame.
-        /// The contents of <paramref name="data"/> should not be modified until this write operation ends.
-        /// </remarks>
-        Task WriteBlockAsync(byte[] data, int startIndex, int length);
-#else
+#if NET35
         /// <summary>
         /// Asynchronously writes a block to the stream.
         /// </summary>
@@ -118,6 +105,23 @@ namespace SharpAvi.Output
         /// </summary>
         /// <seealso cref="BeginWriteBlock"/>
         void EndWriteBlock(IAsyncResult asyncResult);
+#else
+        /// <summary>
+        /// Asynchronously writes a block of audio data.
+        /// </summary>
+        /// <param name="data">Data buffer.</param>
+        /// <param name="startIndex">Start index of data.</param>
+        /// <param name="length">Length of data.</param>
+        /// <returns>
+        /// A task representing the asynchronous write operation.
+        /// </returns>
+        /// <remarks>
+        /// Division of audio data into blocks may be arbitrary.
+        /// However, it is reasonable to write blocks of approximately the same duration
+        /// as a single video frame.
+        /// The contents of <paramref name="data"/> should not be modified until this write operation ends.
+        /// </remarks>
+        Task WriteBlockAsync(byte[] data, int startIndex, int length);
 #endif
 
         /// <summary>
@@ -126,7 +130,7 @@ namespace SharpAvi.Output
         int BlocksWritten { get; }
     }
 
-    
+#if !NET35
     namespace Contracts
     {
         [ContractClassFor(typeof(IAviAudioStream))]
@@ -227,17 +231,7 @@ namespace SharpAvi.Output
                 Contract.Requires(startIndex + length <= data.Length);
             }
 
-#if FX45
-            public Task WriteBlockAsync(byte[] data, int startIndex, int length)
-            {
-                Contract.Requires(data != null);
-                Contract.Requires(startIndex >= 0);
-                Contract.Requires(length >= 0);
-                Contract.Requires(startIndex + length <= data.Length);
-                Contract.Ensures(Contract.Result<Task>() != null);
-                throw new NotImplementedException();
-            }
-#else
+#if NET35
             public IAsyncResult BeginWriteBlock(byte[] data, int startIndex, int length, AsyncCallback userCallback, object stateObject)
             {
                 Contract.Requires(data != null);
@@ -251,6 +245,16 @@ namespace SharpAvi.Output
             public void EndWriteBlock(IAsyncResult asyncResult)
             {
                 Contract.Requires(asyncResult != null);
+                throw new NotImplementedException();
+            }
+#else
+            public Task WriteBlockAsync(byte[] data, int startIndex, int length)
+            {
+                Contract.Requires(data != null);
+                Contract.Requires(startIndex >= 0);
+                Contract.Requires(length >= 0);
+                Contract.Requires(startIndex + length <= data.Length);
+                Contract.Ensures(Contract.Result<Task>() != null);
                 throw new NotImplementedException();
             }
 #endif
@@ -281,4 +285,5 @@ namespace SharpAvi.Output
             }
         }
     }
+#endif
 }
