+++ SharpAvi/Output/IAviVideoStream.cs
@@ -1,6 +1,6 @@
 ﻿using System;
+#if !NET35
 using System.Diagnostics.Contracts;
-#if FX45
 using System.Threading.Tasks;
 #endif
 
@@ -12,7 +12,9 @@ namespace SharpAvi.Output
     /// <remarks>
     /// After the first invocation of <see cref="WriteFrame"/> no properties of the stream can be changed.
     /// </remarks>
+#if !NET35
     [ContractClass(typeof(Contracts.AviVideoStreamContract))]
+#endif
     public interface IAviVideoStream : IAviStream
     {
         /// <summary>Frame width.</summary>
@@ -38,18 +40,7 @@ namespace SharpAvi.Output
         /// <param name="length">Length of the frame data.</param>
         void WriteFrame(bool isKeyFrame, byte[] frameData, int startIndex, int length);
 
-#if FX45
-        /// <summary>Asynchronously writes a frame to the stream.</summary>
-        /// <param name="isKeyFrame">Is this frame a key frame?</param>
-        /// <param name="frameData">Array containing the frame data.</param>
-        /// <param name="startIndex">Index of the first byte of the frame data.</param>
-        /// <param name="length">Length of the frame data.</param>
-        /// <returns>A task that represents the asynchronous write operation.</returns>
-        /// <remarks>
-        /// The contents of <paramref name="frameData"/> should not be modified until this write operation ends.
-        /// </remarks>
-        Task WriteFrameAsync(bool isKeyFrame, byte[] frameData, int startIndex, int length);
-#else
+#if NET35
         /// <summary>
         /// Asynchronously writes a frame to the stream.
         /// </summary>
@@ -71,6 +62,17 @@ namespace SharpAvi.Output
         /// </summary>
         /// <seealso cref="BeginWriteFrame"/>
         void EndWriteFrame(IAsyncResult asyncResult);
+#else
+        /// <summary>Asynchronously writes a frame to the stream.</summary>
+        /// <param name="isKeyFrame">Is this frame a key frame?</param>
+        /// <param name="frameData">Array containing the frame data.</param>
+        /// <param name="startIndex">Index of the first byte of the frame data.</param>
+        /// <param name="length">Length of the frame data.</param>
+        /// <returns>A task that represents the asynchronous write operation.</returns>
+        /// <remarks>
+        /// The contents of <paramref name="frameData"/> should not be modified until this write operation ends.
+        /// </remarks>
+        Task WriteFrameAsync(bool isKeyFrame, byte[] frameData, int startIndex, int length);
 #endif
 
         /// <summary>
@@ -79,7 +81,8 @@ namespace SharpAvi.Output
         int FramesWritten { get; }
     }
 
-    
+
+#if !NET35
     namespace Contracts
     {
         [ContractClassFor(typeof(IAviVideoStream))]
@@ -142,17 +145,7 @@ namespace SharpAvi.Output
                 Contract.Requires(startIndex + length <= frameData.Length);
             }
 
-#if FX45
-            public Task WriteFrameAsync(bool isKeyFrame, byte[] frameData, int startIndex, int length)
-            {
-                Contract.Requires(frameData != null);
-                Contract.Requires(startIndex >= 0);
-                Contract.Requires(length >= 0);
-                Contract.Requires(startIndex + length <= frameData.Length);
-                Contract.Ensures(Contract.Result<Task>() != null);
-                throw new NotImplementedException();
-            }
-#else
+#if NET35
             public IAsyncResult BeginWriteFrame(bool isKeyFrame, byte[] frameData, int startIndex, int length, AsyncCallback userCallback, object stateObject)
             {
                 Contract.Requires(frameData != null);
@@ -166,6 +159,16 @@ namespace SharpAvi.Output
             public void EndWriteFrame(IAsyncResult asyncResult)
             {
                 Contract.Requires(asyncResult != null);
+            }
+#else
+            public Task WriteFrameAsync(bool isKeyFrame, byte[] frameData, int startIndex, int length)
+            {
+                Contract.Requires(frameData != null);
+                Contract.Requires(startIndex >= 0);
+                Contract.Requires(length >= 0);
+                Contract.Requires(startIndex + length <= frameData.Length);
+                Contract.Ensures(Contract.Result<Task>() != null);
+                throw new NotImplementedException();
             }
 #endif
 
@@ -200,4 +203,5 @@ namespace SharpAvi.Output
             }
         }
     }
+#endif
 }
