--- SharpAvi/Properties/AssemblyInfo.cs
@@ -1,16 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("SharpAvi")]
-[assembly: AssemblyDescription("A simple API for creating AVI files")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("SharpAvi")]
-[assembly: AssemblyCopyright("Copyright © Vasili Maslov 2013-2019")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-[assembly: ComVisible(false)]
-
-[assembly: AssemblyVersion("2.1.2.0")]
-[assembly: AssemblyFileVersion("2.1.2.0")]
